streamlit_app/requirements.txt
numpy
pandas
streamlit
Pillow
Matplotlib
Plotly
